能够启动新的应用了，但是ipc有问题

+++ mkrtos_knl/inc/knl/prot.h
@@ -58,7 +58,7 @@ typedef struct msg_tag
     .map_buf_len = (buf_words),                                  \
     .prot = (p)})
 #define msg_tag_get_prot(tag) \
-    ((int16_t)((tag).prot))
+    ((int)((tag).prot))
 #define msg_tag_get_val(tag) msg_tag_get_prot(tag)
 
 typedef union syscall_prot
